[modify] rename ui to dashboard
