+++ dashboard/nodes/test-node.js
@@ -1,10 +1,10 @@
 module.exports = function (RED) {
-  const ui = require("../ui")(RED);
+  const dashboard = require("../dashboard")(RED);
 
   function TestNode(config) {
     const node = this;
     RED.nodes.createNode(node, config);
-    console.log("Test Node created" + ui.testFunc());
+    console.log("Test Node created " + dashboard.testFunc());
   }
 
   RED.nodes.registerType("test-node", TestNode);
